feat: add mock on server
- add mock functions on server side with same typings as client-side
- add gentle warning when using mock during development
- log error instead of throwing (throwing crashes the app)
